* Avoid various overflows in gd_getdata, gd_putdata, gd_seek. * Avoid using "long long" as it's, apparently, not C89 enough.  Instead we use   the C99 [u]int64_t, which is probably less C89, but can at least be used to   work around the problem. * Similarly, avoid "%ll" in printfs, using the PRI* C99 macros instead. * Fix most of the bindings to allow using GD_NULL when requesting data. * PYTHON: Make return_type optional for get_carray_slice and friends.

+++ bindings/f77/fgetdata.c
@@ -151,16 +151,16 @@ static gd_triplet_t _GDF_SetTriplet(gd_windop_t op, const void *data)
     case GD_WINDOP_EQ:
     case GD_WINDOP_NE:
       t.i = *(int32_t*)data;
-      dreturn("%lli", (long long)t.i);
+      dreturn("%" PRId64, t.i);
       break;
     case GD_WINDOP_SET:
     case GD_WINDOP_CLR:
       t.u = *(int32_t*)data;
-      dreturn("%llu", (unsigned long long)t.u);
+      dreturn("%" PRIu64, t.u);
       break;
     default:
       t.r = *(double*)data;
-      dreturn("%g", t.r);
+      dreturn("%.15g", t.r);
       break;
   }
