+++ util/checkdirfile.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 2007-2010, 2012, 2014 D. V. Wiebe
+/* Copyright (C) 2007-2010, 2012, 2014, 2015 D. V. Wiebe
  *
  ***************************************************************************
  *
@@ -248,8 +248,7 @@ int main(int argc, char* argv[])
     return 1;
   }
 
-  printf("  Found %llu %s.\n", (unsigned long long)n,
-        n == 1 ? "frame" : "frames");
+  printf("  Found %" PRIu64 " %s.\n", n, n == 1 ? "frame" : "frames");
 
   gd_close(dirfile);
   return 0;
